Added cast to fix large negative

+++ src/tools/helper/parameters.c
@@ -78,10 +78,10 @@ void set_ice_candidates(
 
         // Set ICE candidate values
         EOR(odict_entry_add(node, "foundation", ODICT_STRING, foundation));
-        EOR(odict_entry_add(node, "priority", ODICT_INT, priority));
+        EOR(odict_entry_add(node, "priority", ODICT_INT, (int64_t)priority));
         EOR(odict_entry_add(node, "ip", ODICT_STRING, ip));
         EOR(odict_entry_add(node, "protocol", ODICT_STRING, rawrtc_ice_protocol_to_str(protocol)));
-        EOR(odict_entry_add(node, "port", ODICT_INT, port));
+        EOR(odict_entry_add(node, "port", ODICT_INT, (int64_t)port));
         EOR(odict_entry_add(node, "type", ODICT_STRING, rawrtc_ice_candidate_type_to_str(type)));
         if (protocol == RAWRTC_ICE_PROTOCOL_TCP) {
             EOR(odict_entry_add(node, "tcpType", ODICT_STRING,
@@ -91,7 +91,7 @@ void set_ice_candidates(
             EOR(odict_entry_add(node, "relatedAddress", ODICT_STRING, related_address));
         }
         if (related_port) {
-            EOR(odict_entry_add(node, "relatedPort", ODICT_INT, related_port));
+            EOR(odict_entry_add(node, "relatedPort", ODICT_INT, (int64_t)related_port));
         }
 
         // Add to array
